requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.7.4
ollama>=0.1.6
httpx>=0.25.2
python-dotenv>=1.0.0
